+++ src/commands/command/codecs/query.ts
@@ -1,0 +1,7 @@
+import { encode } from "../inputs/query.js";
+import { decode } from "../outputs/query.js";
+
+export const QueryCodec = {
+  encode,
+  decode,
+};