+++ src/commands/command/outputs/index.ts
@@ -3,3 +3,4 @@ export type { DeleteItemOutput } from "./delete-item.js";
 export type { DeleteTableOutput } from "./delete-table.js";
 export type { GetItemOutput } from "./get-item.js";
 export type { PutItemOutput } from "./put-item.js";
+export type { QueryOutput } from "./query.js";