+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.58.0] - 2026-02-20
+
+### Added
+
+- The `Query` command.
+
+### Changed
+
+- Renamed `QueryParams` to `QueryInput` and `QueryResult` to `QueryOutput`.
+- Reordered the type parameters `<K, T>` from the `query` APIs to be `<T, K>` first.
+We put `T` first, because can easily be inferred. Therefore, most client code might
+only want to cast the return type.
+
 ## [0.57.0] - 2026-02-19
 
 ### Added
@@ -733,6 +746,7 @@ intuitive.
 - Initial release of the package! Move the implementation work in progress from another
 project to here.
 
+[0.58.0]: https://github.com/infra-blocks/ts-aws-dynamodb/compare/v0.57.0...v0.58.0
 [0.57.0]: https://github.com/infra-blocks/ts-aws-dynamodb/compare/v0.56.0...v0.57.0
 [0.56.0]: https://github.com/infra-blocks/ts-aws-dynamodb/compare/v0.55.0...v0.56.0
 [0.55.0]: https://github.com/infra-blocks/ts-aws-dynamodb/compare/v0.54.0...v0.55.0